refactor preview panel structure and helpers
