+++ src/features/preview/ui/panel/DeletePageAction.tsx
@@ -1,0 +1,66 @@
+import type { ReactNode } from "react";
+import { Trash2 } from "lucide-react";
+
+import { Button } from "@/components/ui/button";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+
+interface DeletePageActionProps {
+  title: string;
+  description: string;
+  disabled?: boolean;
+  onDelete: () => void;
+  children?: ReactNode;
+}
+
+export function DeletePageAction({
+  title,
+  description,
+  disabled,
+  onDelete,
+  children,
+}: DeletePageActionProps) {
+  return (
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        {children ?? (
+          <Button
+            size="icon"
+            variant="outline"
+            disabled={disabled}
+            className="h-8 w-8 text-destructive"
+          >
+            <Trash2 className="h-3.5 w-3.5" />
+          </Button>
+        )}
+      </AlertDialogTrigger>
+
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>{title}</AlertDialogTitle>
+          <AlertDialogDescription>{description}</AlertDialogDescription>
+        </AlertDialogHeader>
+
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+
+          <AlertDialogAction
+            onClick={onDelete}
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+          >
+            Delete Page
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
